fix: QuestionDimension.weight must map to BigDecimal, not Double
Hibernate schema validation failed at startup: question_dimensions.weight
is DECIMAL(4,3) in the DB (V30) but the entity mapped it as Double, which
Hibernate expects as DOUBLE PRECISION -- "wrong column type encountered in
column [weight] ... found [decimal], but expecting [double precision]".

Switch the entity field to BigDecimal (precision=4, scale=3) and convert
at the two read/write boundaries that need a primitive (scoring engine's
weightsFor(), QuestionServiceImpl's persistDimensionWeights()). Scoring
math, validation rules, and the DB schema are unchanged -- QuestionRequest
still carries plain Double from the API, converted to BigDecimal only when
persisted.

Verified against a live DB: Flyway validates 30 migrations with no
re-attempted migration, and Hibernate's EntityManagerFactory initializes
without error.

+++ exam-portal-backend/src/main/java/com/project/examportalbackend/services/implementation/PsychometricReportServiceImpl.java
@@ -237,7 +237,7 @@ public class PsychometricReportServiceImpl implements PsychometricReportService 
         Map<String, Double> weights = new HashMap<>();
         for (com.project.examportalbackend.models.QuestionDimension qd : questionDimensionRepository.findByQuesId(quesId)) {
             if (qd.getWeight() != null) {
-                weights.put(qd.getDimensionCode(), qd.getWeight());
+                weights.put(qd.getDimensionCode(), qd.getWeight().doubleValue());
             }
         }
         return weights;
